fixed test names
