+++ src/tests/smoke/both/auth/sign-in.spec.ts
