+++ src/tests/smoke/card-dashboard.spec.ts
