+++ src/tests/smoke/desktop/demo/discuss-stage.spec.ts
