+++ src/tests/smoke/desktop/demo/group-stage.spec.ts
