+++ src/tests/smoke/desktop/demo/vote-stage.spec.ts
