+++ src/tests/smoke/mobile/auth/sign-out.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
-import { SignOutOnMobile } from 'src/common-events/auth/logout-for/mobile/mobile';
-import { LogInAsUserOne } from 'src/common-events/auth/login-as/user-one';
+import { SignOutOnMobileBy } from '../../../../common-events/auth/logout-for/mobile';
+import { LogInAsUserOneBy } from '../../../../common-events/auth/login-as/user-one';
 import 'dotenv/config';
 
 const base_url = process.env.BASE_URL;
