+++ src/tests/smoke/mobile/demo/discuss-stage.spec.ts
