+++ src/tests/smoke/mobile/demo/group-stage.spec.ts
