+++ src/tests/smoke/mobile/demo/vote-stage.spec.ts
