+++ src/tests/smoke/sprint-poker.spec.ts
