+++ src/tests/smoke/upgrade-to-pro.spec.ts
